chore: api 컨벤션 정리
