--- src/apis/note/deleteNote.ts
@@ -1,4 +1,0 @@
-import instance from "../apiClient";
-
-export const deleteNote = async (noteId: number) =>
-  await instance.delete(`/notes/${noteId}`);
