+++ src/apis/noteApi.ts
@@ -1,10 +1,17 @@
 import {
   CreateNoteBodyDto,
+  TeamIdNotesGet200ResponseNotesInner,
+  teamIdNotesGetParams,
   TeamIdNotesPost201Response,
   UpdateNoteBodyDto,
 } from "@/types/types";
 
 import instance from "./apiClient";
+
+interface FetchNotesParams {
+  pageParam?: number;
+  goalId: number;
+}
 
 const noteApi = {
   createNote: async (
@@ -18,10 +25,34 @@ const noteApi = {
   ): Promise<TeamIdNotesPost201Response> => {
     return (await instance.patch(`/notes/${noteId}`, data)).data;
   },
+
+  fetchNotes: async ({ pageParam, goalId }: FetchNotesParams) => {
+    const PAGE_SIZE = 6;
+
+    const params: teamIdNotesGetParams = {
+      goalId,
+      cursor: pageParam,
+      size: PAGE_SIZE,
+    };
+    return (await instance.get("notes", { params })).data;
+  },
+
   fetchNote: async (noteId: number): Promise<TeamIdNotesPost201Response> => {
     const { data } = await instance.get(`/notes/${noteId}`);
 
     return data;
+  },
+
+  fetchNoteDetail: async (
+    noteId: number,
+  ): Promise<TeamIdNotesGet200ResponseNotesInner | null> => {
+    if (noteId == null) return Promise.resolve(null);
+
+    return (await instance.get(`/notes/${noteId}`)).data;
+  },
+
+  deleteNote: async (noteId: number) => {
+    return await instance.delete(`/notes/${noteId}`);
   },
 };
